requirements.txt
python-dotenv==1.0.1
boto3>=1.35.0
botocore>=1.35.0
s3transfer>=0.10.0